finished Usuário

+++ CSI477-2019-02-Atividade-PHP/protocolos/app/Http/Controllers/SubjectController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use DB;
 use App\Subject;
 use Illuminate\Http\Request;
 
@@ -25,7 +26,7 @@ class SubjectController extends Controller
      */
     public function create()
     {
-        //
+        return view('subject.create');
     }
 
     /**
@@ -36,7 +37,7 @@ class SubjectController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        dd($request);
     }
 
     /**
@@ -58,7 +59,8 @@ class SubjectController extends Controller
      */
     public function edit(Subject $subject)
     {
-        //
+        dd($subject);
+        return view('subject.edit', compact('subject'));
     }
 
     /**
@@ -82,7 +84,12 @@ class SubjectController extends Controller
     public function destroy(Request $req)
     {
         $cert = Subject::find($req->query('subject'));
-        $cert->delete();
-        return redirect()->route('subjects')->withStatus(__('Cartificado Deletado Com Sucesso'));
+        $requests = DB::table('requests')->where('requests.subject_id', '=', $req->query('subject'))->get();
+        if(count($requests) > 0){
+            return redirect()->route('subjects')->withErrors('Certificado Possui Requerimentos Associados');
+        }else{
+            $cert->delete();
+            return redirect()->route('subjects')->withStatus(__('Cartificado Deletado Com Sucesso'));
+        }
     }
 }
